refactor(engine): one idea per file, and scope the docs to the engine
Docs first: the ADR and traps notes read as repo-wide policy at the root, when
they only bind the engine — a learning track that may never ship. Moved to
packages/engine/docs/ alongside the plan, so nothing implies mentis@0.2.x is
affected.

Then split every file down to a single idea. The two worst offenders were
replay.ts (330 lines) and event-log.ts (290); the largest file is now 141.
Two of the splits are structural rather than cosmetic:

- log/describe-event.ts separates what gets *reported* about an event from the
  listener plumbing in create-event-log.ts. Those describe functions are plain
  event -> {summary, detail}, so they are now readable and testable without
  attaching to a live editor — which is where the six new tests came from.
- replay/parse-script.ts holds no browser code at all, so it stays in the pure
  vitest project and the Playwright suite can share the script syntax.

Also replaces the literal invisible characters in the whitespace map with
String.fromCodePoint. They were unreviewable, and two look-alikes had silently
collided into a duplicate object key that only surfaced as a typecheck error
during the split — exactly the failure the map exists to make visible.

Adds a root CLAUDE.md (un-ignored so it ships with the repo): 21 bullets, all
constraints or footguns, nothing a reader could get from package.json.

51 tests, up from 41. Typecheck clean; every module verified to resolve
through the dev server.
