+++ CLAUDE.md
@@ -1,0 +1,23 @@
+# mentis — a React mention tagger that hooks into existing inputs, plus a greenfield contenteditable engine intended to replace its internals
+
+- **`packages/mentis` is frozen at 0.2.x**: bug fixes and test coverage only, no new features. Don't add capability there to close a gap `packages/engine` is meant to close.
+- **`packages/engine` is a learning project, optimised for interesting over shipping.** When a decision is ambiguous, pick the option that teaches more. Never introduce migration deadlines, backwards-compatibility pressure, or release urgency there. It is `private: true` — keep it that way. Context in `packages/engine/docs/plan.md`.
+- **One idea per file.** A file that needs "and" to describe it is two files. Split rather than grow — this codebase optimises for reviewability over locality.
+- **Prefer plain exported functions over classes**, and keep mutable state in a single wiring module at the edge (`create-*.ts`) rather than spread through the logic it coordinates. `log/describe-event.ts` vs `log/create-event-log.ts` is the pattern: pure description separated from listener plumbing.
+- **Nothing below `adapters/` in the engine may import a framework.** This is the whole reason the eventual React/Vue/vanilla wrappers can be ~100 lines each.
+- **happy-dom is not a source of truth for caret, `Selection`, `Range`, `beforeinput`, or composition.** Assert structure and "does it throw" there, never caret semantics. Commit `a6fcfd0` bent production code to satisfy the fake DOM — if a test needs the environment bent, the test is at the wrong layer.
+- **Caret position, native editing, undo, and clipboard behaviour are tested in Playwright** (`e2e/`), never vitest.
+- When asked to **"add an e2e test to prevent this in future"**, copy `e2e/regressions/_template.spec.ts` to `e2e/regressions/<symptom-slug>.spec.ts` — named after the symptom, not the cause. One bug per file, reproduced through user actions only, never by importing from `packages/mentis/src`. If the bug doesn't involve caret, native editing, or clipboard, push it down to a unit or component test and say that's what you did.
+- **Never compare a character against `" "`.** A non-breaking space (U+00A0) renders identically but fails `===`, while `/\s/` matches it — so regex checks pass by accident and equality checks fail silently. Normalise once at the model boundary, not per call site.
+- **Never measure caret offsets with `Range.toString()`.** It returns only `Text` data and ignores `<br>`, so every offset after a line break is off by one. Use `textLength`; a `<br>` is exactly one `\n`, per `packages/engine/docs/adr/0001-line-breaks-as-newline-characters.md`.
+- **Never type invisible characters literally into source.** Name them by code point (`String.fromCodePoint(0x00a0)`). Two look-alike invisible characters silently become a duplicate object key, and no reviewer can see the difference.
+- **Script-created events cannot cause editing** — they are `isTrusted: false`, so the browser skips the default action. Use `document.execCommand` when you need a genuine `beforeinput`/`input` pair on the native undo stack, and accept that IME and mobile autocorrect can only be exercised by hand or through Playwright/CDP.
+- **In `packages/mentis` the DOM *is* the state** — `extractMentionData` re-derives truth from it every keystroke, so any DOM quirk is a correctness bug. Expect fixes there to be whack-a-mole; that is the reason the engine exists.
+- **Add a changeset for any user-facing `mentis` change.** Releases publish from the changesets workflow on merge to main — never bump a version or publish by hand.
+- **`packages/mentis/README.md` and `LICENSE` are gitignored deliberately**: `scripts/prepack.sh` copies them in at pack time. Don't create them or "fix" the gitignore.
+- **CI only runs `--filter mentis`.** `packages/engine` is not built, typechecked, or tested there, so run `pnpm --filter @mentis/engine typecheck` and `... test` locally before committing to it.
+- **`pnpm --filter mentis test` watches on a TTY.** Use `vitest run` when you want a single pass; `@mentis/engine` already does.
+- **Work happens in parallel git worktrees** — `../mentis-engine` on `feat/engine`, `../mentis-e2e` on `test/e2e-harness`. Each needs its own `pnpm install`; they don't share `node_modules`.
+- **Don't rebase or revive `overhaul-html-node-logic`.** It's an archived design sketch, 11 months stale, with a conflicting `extractMentionData` rewrite on main. Only its `v2/mention-query/*` logic and tests are worth salvaging.
+- **Record decisions, not facts.** A commitment future work inherits goes in `packages/engine/docs/adr/` with an explicit revisit-when trigger; a platform trap that will bite again goes in `packages/engine/docs/notes/contenteditable-traps.md`; a local "don't break this" goes in a comment at the line. The test is whether someone returning after a month away would find it.
+- **Done means typecheck and tests green before the commit, and green CI on the PR.** State plainly what is still unverified — browser-only behaviour especially, since no local check covers it.
