+++ packages/engine/dev/need.ts
@@ -1,0 +1,6 @@
+/** Query an element the harness page is required to contain, failing loudly if not. */
+export const need = <T extends HTMLElement>(selector: string): T => {
+  const element = document.querySelector<T>(selector);
+  if (!element) throw new Error(`Missing element: ${selector}`);
+  return element;
+};
