+++ packages/engine/src/devtools/text/escape-html.ts
@@ -1,0 +1,7 @@
+/** Ampersand first, so entities produced here are not double-escaped. */
+export const escapeHtml = (text: string): string =>
+  text
+    .replace(/&/g, "&amp;")
+    .replace(/</g, "&lt;")
+    .replace(/>/g, "&gt;")
+    .replace(/"/g, "&quot;");
